changelog, version, removed utils_check, ArrayType, doc
