+++ doc/entrypoints/correction.rst
@@ -6,3 +6,6 @@ Correction
 
 .. automodule:: omc3.response_creator
     :members:
+
+.. automodule:: omc3.check_corrections
+    :members:
