+++ doc/entrypoints/plotting.rst
@@ -10,6 +10,9 @@ Plotting
 .. automodule:: omc3.plotting.plot_spectrum
     :members:
 
+.. automodule:: omc3.plotting.plot_checked_corrections
+    :members:
+
 .. automodule:: omc3.plotting.plot_amplitude_detuning
     :members:
